Split out toolboxes by type
