+++ src/editor/editor.ts
@@ -24,7 +24,10 @@ import * as Blockly from 'blockly/core';
 import { extendedPythonGenerator } from './extended_python_generator';
 import { GeneratorContext } from './generator_context';
 import * as commonStorage from '../storage/common_storage';
-import { getToolboxJSON } from '../toolbox/toolbox';
+import * as toolboxOpmode from '../toolbox/toolbox_opmode';
+import * as toolboxMechanism from '../toolbox/toolbox_mechanism';
+import * as toolboxRobot from '../toolbox/toolbox_robot';
+
 //import { testAllBlocksInToolbox } from '../toolbox/toolbox_tests';
 import { MethodsCategory} from '../toolbox/methods_category';
 
@@ -182,13 +185,6 @@ export class Editor {
 
   public updateToolbox(shownPythonToolboxCategories: Set<string>): void {
     if (this.currentModule) {
-      if (this.currentModule.moduleType === commonStorage.MODULE_TYPE_PROJECT) {
-        // If we are editing a Project, we don't add any additional blocks to
-        // the toolbox.
-        this.setToolbox(getToolboxJSON([], shownPythonToolboxCategories));
-        return;
-      }
-      // Otherwise, we add the exported blocks from the Project.
       if (!this.projectContent) {
         // The Project content hasn't been fetched yet. Try again in a bit.
         setTimeout(() => {
@@ -196,9 +192,21 @@ export class Editor {
         }, 50);
         return;
       }
+      switch(this.currentModule.moduleType){
+        case commonStorage.MODULE_TYPE_PROJECT:
+          this.setToolbox(toolboxRobot.getToolboxJSON(shownPythonToolboxCategories));
+          break;
+        case commonStorage.MODULE_TYPE_MECHANISM:
+          this.setToolbox(toolboxMechanism.getToolboxJSON(shownPythonToolboxCategories));
+          break;
+        case commonStorage.MODULE_TYPE_OPMODE:
+/*
       const robotBlocks = commonStorage.extractExportedBlocks(
          this.currentModule.projectName, this.projectContent);
-      this.setToolbox(getToolboxJSON(robotBlocks, shownPythonToolboxCategories));
+*/
+          this.setToolbox(toolboxOpmode.getToolboxJSON(shownPythonToolboxCategories));
+          break;
+      }
     }
   }
